requirements.txt
streamlit
sqlalchemy
pandas
numpy
Pillow
opencv-python-headless
qrcode
psycopg2-binary
st-supabase-connection
streamlit
sqlalchemy
pandas
numpy
Pillow
opencv-python-headless
qrcode
